mISDNuser-alignment-management.patch
ARM silently misreads misaligned multibyte objects -- so if memory at 0 had

0: 11 22 33 44 55 66

Then an int pointer to 0x0 will read 0x44332211 as it should, but an int
pointer to 0x1 will act differently according to the arch being compiled

x86: 0x55443322
ARM: 0x11443322

and there will be NO DIAGNOSTIC.  A typical case where the problem is
introduced is to allocate a char array, which has random alignment, and then
cast that pointer to be pointing to a struct.  The whole of the struct is then
fatally misaligned.

This patch adds some defines that allow simple checking of pointer alignment,
but more importantly to enforce int alignment of buffer allocation from the
stack.

Signed-off-by: Andy Green <andy@warmcat.com>